feat: [PL-30349]: Moved code to correct service class
