+++ audit-event-streaming/service/src/main/java/io/harness/auditevent/streaming/services/StreamingBatchService.java
@@ -7,6 +7,7 @@
 
 package io.harness.auditevent.streaming.services;
 
+import io.harness.audit.entities.streaming.StreamingDestination;
 import io.harness.auditevent.streaming.entities.BatchStatus;
 import io.harness.auditevent.streaming.entities.StreamingBatch;
 
@@ -16,6 +17,20 @@ import java.util.Optional;
 public interface StreamingBatchService {
   Optional<StreamingBatch> get(
       String accountIdentifier, String streamingDestinationIdentifier, List<BatchStatus> batchStatusList);
+
   Optional<StreamingBatch> getLatest(String accountIdentifier, String streamingDestinationIdentifier);
+
   StreamingBatch update(String accountIdentifier, StreamingBatch streamingBatch);
+
+  /**
+   * Returns the last streaming batch for the given streamingDestination. If no streaming is batch found then it creates
+   * a new batch and returns it. If last batch was a {@link BatchStatus#SUCCESS SUCCESS} then it creates next batch
+   * based on previous batch and returns it. If last batch {@link BatchStatus#FAILED FAILED} then it increments the
+   * retry count and return the batch.
+   * @param streamingDestination {@link StreamingDestination StreamingDestination} for which batch is to be returned
+   * @param timestamp Uses this timestamp as {@link StreamingBatch#setEndTime(Long) endTime} when creating a new {@link
+   *     StreamingBatch StreamingBatch}
+   * @return {@link StreamingBatch StreamingBatch}
+   */
+  StreamingBatch getLastStreamingBatch(StreamingDestination streamingDestination, Long timestamp);
 }
